ci: fix molecule CI (fedora35)
by using https://github.com/search?q=fedora35+user%3Ageerlingguy&type=code&ref=advsearch
i found one geerlingguy repository with CI still passing and fedora35 being in the list of tested containers. it had the following difference which this commit replicates
https://github.com/geerlingguy/ansible-role-containerd/commit/07ced7cca8db589c6b2e031b687565c61ff951a7

i tested it in a tmate session and it worked using this change!
committing this now to a PR to see if it does damage to other distro's or if it's OK
(i am really not a docker or linux internals expert, still need to read my books i postponed for so long for advancing in networking-side career :))